#58: unify emitter dispatch — extract EmitPython module
Resolve the historical asymmetry where Python was emitted by an inline
private method in Emit.scala while Go/TS were separate modules. Move the
Python emitter + its private helpers and template-context case classes to
specrest.codegen.python.EmitPython (mirroring specrest.codegen.go.EmitGo /
specrest.codegen.ts.EmitTs). Emit.scala is now purely the shared
EmittedFile / EmitOptions types plus a uniform language dispatcher.

Pure structural move, behavior-neutral: codegen suite green (183/183),
downstream modules compile.
